bug(search): solve advanced search fields order
solves advanced search fields on the homepage's home search form
showing behind cities section
